+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# HEAD
+
+- Fixed: `dollar-variable-colon-newline-after` now allows multiline variables when `always` option is used.
+
 # 2.3.0
 
 - Added: `dollar-variable-default` rule.